spawner, spatial hashing, etc

+++ World Of Rectangle/Game/Global.cs
@@ -9,7 +9,7 @@ namespace World_Of_Rectangle.Game
 {
     class Global
     {
-        public const float TILE_SIZE = 64;
+        public const int TILE_SIZE = 64;
         public static Vector3 scaleValues;
         public static readonly Vector2 REFERENCE_SCREENSIZE = new Vector2(1408, 832);
         public static Vector2 SCREENSIZE;
